Module carapax::access

Available on crate feature access only.

Access control

Structs

AccessPredicate

An access predicate for protecting a Handler with a specified AccessPolicy.

AccessRule

Represents an access rule containing information about a principal and access grant status.

InMemoryAccessPolicy

In-memory access policy implementation.

Enums

Principal

Represents a principal entity that decides whether an crate::access::AccessRule should accept an update.

Traits

AccessExt

Provides a shortcut for wrapping a Handler by an AccessPredicate.

AccessPolicy

Decides whether HandlerInput should be processed or not.