Struct carapax::prelude::AnswerInlineQuery

pub struct AnswerInlineQuery { /* fields omitted */ }

Use this method to send answers to an inline query

No more than 50 results per query are allowed

Methods

impl AnswerInlineQuery

pub fn new<S>(
inline_query_id: S,
results: Vec<InlineQueryResult>
) -> AnswerInlineQuery where
S: Into<String>,

Creates a new AnswerInlineQuery with empty optional parameters

Arguments

inline_query_id - Unique identifier for the answered query

results - An array of results for the inline query

pub fn cache_time(self, cache_time: i64) -> AnswerInlineQuery

Maximum amount of time in seconds that the result of the inline query may be cached on the server

Defaults to 300

pub fn personal(self, is_personal: bool) -> AnswerInlineQuery

Cache results on the server side only for the user that sent the query

By default, results may be returned to any user who sends the same query

pub fn next_offset<S>(self, next_offset: S) -> AnswerInlineQuery where
S: Into<String>,

Offset that a clien should send in the next query with the same text to receive more results

Pass an empty string if there are no more results or if you don‘t support pagination Offset length can’t exceed 64 bytes

pub fn switch_pm_text<S>(self, switch_pm_text: S) -> AnswerInlineQuery where
S: Into<String>,

Clients will display a button with specified text that switches the user to a private chat with the bot and sends the bot a start message with the parameter switch_pm_parameter

pub fn switch_pm_parameter<S>(self, switch_pm_parameter: S) -> AnswerInlineQuery where
S: Into<String>,

Deep-linking parameter for the /start message sent to the bot when user presses the switch button

1-64 characters, only A-Z, a-z, 0-9, _ and - are allowed

Example: An inline bot that sends YouTube videos can ask the user to connect the bot to their YouTube account to adapt search results accordingly To do this, it displays a ‘Connect your YouTube account’ button above the results, or even before showing any The user presses the button, switches to a private chat with the bot and, in doing so, passes a start parameter that instructs the bot to return an oauth link Once done, the bot can offer a switch_inline button so that the user can easily return to the chat where they wanted to use the bot's inline capabilities