Trait carapax::prelude::Handler

pub trait Handler {
    type Input: FromUpdate;
    type Output: Into<HandlerFuture>;
    fn handle(&self, context: &mut Context, input: Self::Input) -> Self::Output;
}

An update handler

Allows to handle specific kinds of update or update itself

Associated Types

type Input: FromUpdate

An object to handle (update, message, inline query, etc...)

See FromUpdate for more information

type Output: Into<HandlerFuture>

A result or future to return

See HandlerFuture for more information

Loading content...
Required methods

fn handle(&self, context: &mut Context, input: Self::Input) -> Self::Output

Handles an object

Arguments

context - A handler context

input - An object obtained from update (update itself, message, etc...)

The context is valid for all handlers that accepts given update

When processing a subsequent update, the old context will be destroyed and replaced with a new one

Loading content...