Struct carapax::ratelimit::KeyedRateLimitPredicate

pub struct KeyedRateLimitPredicate<K, J, M>
where
    K: Key,
{ /* private fields */ }

Available on crate feature ratelimit only.

A predicate with keyed rate limiter.

Each update will have it’s own rate limit under key K.

Implementations

impl<K, J, M> KeyedRateLimitPredicate<K, J, M>
where K: Key,

pub fn with_key<T: Into<K>>(self, key: T) -> Self

Use this method when you need to run a predicate only for a specific key.

If this method is not called, predicate will run for all updates.

Arguments

key - A key to filter by.

impl<K> KeyedRateLimitPredicate<K, NoJitter, MethodDiscard>
where K: Key,

pub fn discard(quota: Quota) -> Self

Creates a new KeyedRateLimitPredicate with the discard method.

Predicate will stop update propagation when the rate limit is reached.

Arguments

quota - A rate limiting quota.

impl<K> KeyedRateLimitPredicate<K, NoJitter, MethodWait>
where K: Key,

pub fn wait(quota: Quota) -> Self

Creates a new KeyedRateLimitPredicate with wait method.

Predicate will pause update propagation when the rate limit is reached.

Arguments

quota - A rate limiting quota.

impl<K> KeyedRateLimitPredicate<K, Jitter, MethodWait>
where K: Key,

pub fn wait_with_jitter(quota: Quota, jitter: Jitter) -> Self

Creates a new KeyedRateLimitPredicate with wait method and jitter.

Predicate will pause update propagation when the rate limit is reached.

Arguments

quota - A rate limiting quota.

jitter - An interval specification for deviating from the nominal wait time.