Struct carapax::Chain

pub struct Chain { /* private fields */ }

Represents a chain of handlers.

The chain allows you to configure multiple handlers for the crate::App.

There are two strategies to run handlers:

Self::once - the chain runs only a first found handler.

Self::all - the chain runs all found handlers.

A Handler considered found when TryFromInput::try_from_input for the handler returns Some.

Handlers are dispatched in the order they are added.

If a handler returns an error, all subsequent handlers will not run.

Implementations

impl Chain

pub fn once() -> Self

Creates a new Chain that runs only the first found handler.

pub fn all() -> Self

Creates a new Chain that runs all given handlers.

pub fn with<H, I, O>(self, handler: H) -> Self
where H: Handler<I, Output = O> + Sync + Clone + 'static, I: TryFromInput + Sync + 'static, O: Into<ChainResult>,

Adds a handler to the chain.

Arguments

handler - The handler to add.

Panics

Panics when trying to add a handler to a shared chain.