carapax

Trait Handler

pub trait Handler<I>: Clone + Send
where
    I: TryFromInput,
{
    type Output: Send;

    // Required method
    fn handle(&self, input: I) -> impl Future<Output = Self::Output> + Send;
}

Allows to handle a specific HandlerInput.

Required Associated Types

type Output: Send

A future output returned by Self::handle method.

Use HandlerResult or any type that can be converted into it if you want to use the handler in crate::App.

It is possible to use any other type, e.g. if you want to use it in a decorator. But finally you need to convert it into HandlerResult.

Required Methods

fn handle(&self, input: I) -> impl Future<Output = Self::Output> + Send

Handles a specific input.

Arguments

input - The input to handle.

See TryFromInput trait implementations for a list of supported types.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.