carapax

Trait PredicateExt

pub trait PredicateExt<P, PI, HI>: Sized {
    // Provided method
    fn with_predicate(self, predicate: P) -> Predicate<P, PI, Self, HI> { ... }
}

Provides a shortcut for wrapping a Handler by a Predicate.

Provided Methods

fn with_predicate(self, predicate: P) -> Predicate<P, PI, Self, HI>

Shortcut to create a wrap a Handler with a Predicate.

Example: handler.predicate(predicate).

Arguments

predicate - A predicate handler.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.