Trait carapax::TryFromUpdate

pub trait TryFromUpdate: Sized {
    type Error: Error + Send + Sync;
    fn try_from_update(update: Update) -> Result<Option<Self>, Self::Error>;
}

Allows to create an input for a handler from given update

Associated Types

type Error: Error + Send + Sync

An error when converting update

Loading content...
Required methods

fn try_from_update(update: Update) -> Result<Option<Self>, Self::Error>

Returns a handler input

Handler will not run if None or Error returned

Loading content...