carapax::types

Enum BackgroundType

pub enum BackgroundType {
    ChatTheme {
        theme_name: String,
    },
    Fill {
        dark_theme_dimming: i64,
        fill: BackgroundFill,
    },
    Pattern {
        document: Document,
        fill: BackgroundFill,
        intensity: i64,
        is_inverted: Option<bool>,
        is_moving: Option<bool>,
    },
    Wallpaper {
        dark_theme_dimming: i64,
        document: Document,
        is_blurred: Option<bool>,
        is_moving: Option<bool>,
    },
}

This object describes the type of a background.

Variants

ChatTheme

The background is taken directly from a built-in chat theme.

Fields

theme_name: String

Name of the chat theme, which is usually an emoji.

Fill

The background is automatically filled based on the selected colors.

Fields

dark_theme_dimming: i64

Dimming of the background in dark themes, as a percentage; 0-100.

fill: BackgroundFill

The background fill.

Pattern

The background is a PNG or TGV (gzipped subset of SVG with MIME type โapplication/x-tgwallpatternโ) pattern to be combined with the background fill chosen by the user.

Fields

document: Document

Document with the pattern.

fill: BackgroundFill

The background fill that is combined with the pattern.

intensity: i64

Intensity of the pattern when it is shown above the filled background; 0-100.

is_inverted: Option<bool>

Whether the background fill must be applied only to the pattern itself. All other pixels are black in this case. For dark themes only.

is_moving: Option<bool>

Whether the background moves slightly when the device is tilted.

Wallpaper

The background is a wallpaper in the JPEG format.

Fields

dark_theme_dimming: i64

Dimming of the background in dark themes, as a percentage; 0-100.

document: Document

Document with the wallpaper.

is_blurred: Option<bool>

Whether the wallpaper is downscaled to fit in a 450x450 square and then box-blurred with radius 12.

is_moving: Option<bool>

Whether the background moves slightly when the device is tilted.