Enum carapax::types::TextEntity

pub enum TextEntity {
    Bold(TextEntityPosition),
    BotCommand(TextEntityPosition),
    Cashtag(TextEntityPosition),
    Code(TextEntityPosition),
    Email(TextEntityPosition),
    Hashtag(TextEntityPosition),
    Italic(TextEntityPosition),
    Mention(TextEntityPosition),
    PhoneNumber(TextEntityPosition),
    Pre {
        position: TextEntityPosition,
        language: Option<String>,
    },
    Strikethrough(TextEntityPosition),
    TextLink {
        position: TextEntityPosition,
        url: String,
    },
    TextMention {
        position: TextEntityPosition,
        user: User,
    },
    Underline(TextEntityPosition),
    Url(TextEntityPosition),
}

Respresents an entity in a text

Variants

Bold(TextEntityPosition)

Bold text

BotCommand(TextEntityPosition)

Bot command

Cashtag(TextEntityPosition)

Cashtag

Code(TextEntityPosition)

Monowidth string

Email(TextEntityPosition)

E-Mail

Hashtag(TextEntityPosition)

Hashtag

Italic(TextEntityPosition)

Italic text

Mention(TextEntityPosition)

User mention (e.g. @username)

PhoneNumber(TextEntityPosition)

Phone number

Pre

Monowidth block

Fields of Pre

position: TextEntityPosition

Position of entity in text

language: Option<String>

Name of the programming language

Strikethrough(TextEntityPosition)

Strikethrough text

TextLink

Clickable text URLs

Fields of TextLink

position: TextEntityPosition

Position of entity in text

url: String

URL that will be opened after user taps on the text

TextMention

Mention user without username

Fields of TextMention

position: TextEntityPosition

Position of entity in text

user: User

Mentioned user

Underline(TextEntityPosition)

Underlined text

Url(TextEntityPosition)

URL

Implementations

impl TextEntity

pub fn bold(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn bot_command(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn cashtag(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn code(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn email(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn hashtag(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn italic(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn mention(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn phone_number(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn strikethrough(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn underline(offset: i64, length: i64) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

pub fn pre<L>(offset: i64, length: i64, language: Option<L>) -> TextEntity where
L: Into<String>,

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

language - The programming language of the entity text

pub fn text_link<U>(offset: i64, length: i64, url: U) -> TextEntity where
U: Into<String>,

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

url - URL that will be opened after user taps on the text

pub fn text_mention(offset: i64, length: i64, user: User) -> TextEntity

Creates a new TextEntity

Arguments

offset - Offset in UTF-16 code units to the start of the entity

length - Length of the entity in UTF-16 code units

user - User to be mentioned