carapax::types

Struct ChatFullInfo

pub struct ChatFullInfo {

    pub id: i64,
    pub chat_type: ChatFullInfoType,
    pub accent_color: AccentColor,
    pub max_reaction_count: i64,
    pub available_reactions: Option<Vec<ReactionType>>,
    pub background_custom_emoji_id: Option<String>,
    pub has_hidden_members: Option<bool>,
    pub has_protected_content: Option<bool>,
    pub has_visible_history: Option<bool>,
    pub message_auto_delete_time: Option<i64>,
    pub photo: Option<ChatPhoto>,
    pub pinned_message: Option<Message>,
    pub profile_accent_color: Option<ProfileAccentColor>,
    pub profile_background_custom_emoji_id: Option<String>,
    pub active_usernames: Option<Vec<String>>,
    pub username: Option<String>,
    pub description: Option<String>,
    pub invite_link: Option<String>,
    pub title: Option<String>,
    pub can_send_paid_media: Option<bool>,
    pub linked_chat_id: Option<i64>,
    pub bio: Option<String>,
    pub birthdate: Option<Birthdate>,
    pub business_intro: Option<BusinessIntro>,
    pub business_location: Option<BusinessLocation>,
    pub business_opening_hours: Option<BusinessOpeningHours>,
    pub emoji_status_custom_emoji_id: Option<String>,
    pub emoji_status_expiration_date: Option<i64>,
    pub first_name: Option<String>,
    pub has_private_forwards: Option<bool>,
    pub has_restricted_voice_and_video_messages: Option<bool>,
    pub last_name: Option<String>,
    pub personal_chat: Option<Chat>,
    pub can_set_sticker_set: Option<bool>,
    pub custom_emoji_sticker_set_name: Option<String>,
    pub has_aggressive_anti_spam_enabled: Option<bool>,
    pub is_direct_messages: Option<bool>,
    pub is_forum: Option<bool>,
    pub join_to_send_messages: Option<bool>,
    pub join_by_request: Option<bool>,
    pub location: Option<ChatLocation>,
    pub parent_chat: Option<Chat>,
    pub slow_mode_delay: Option<i64>,
    pub sticker_set_name: Option<String>,
    pub unrestrict_boost_count: Option<i64>,
    pub permissions: Option<ChatPermissions>,
    pub accepted_gift_types: Option<AcceptedGiftTypes>,

}

Contains full information about a chat.

Fields
id: i64

Unique identifier for this chat.

chat_type: ChatFullInfoType

Type of the chat.

accent_color: AccentColor

Identifier of the accent color for the chat name and backgrounds of the chat photo, reply header, and link preview.

max_reaction_count: i64

The maximum number of reactions that can be set on a message in the chat.

available_reactions: Option<Vec<ReactionType>>

List of available reactions allowed in the chat. If omitted, then all emoji reactions are allowed.

background_custom_emoji_id: Option<String>

Custom emoji identifier of the emoji chosen by the chat for the reply header and link preview background.

has_hidden_members: Option<bool>

Whether non-administrators can only get the list of bots and administrators in the chat.

has_protected_content: Option<bool>

Whether messages from the chat can’t be forwarded to other chats.

has_visible_history: Option<bool>

Whether new chat members will have access to old messages; available only to chat administrators.

message_auto_delete_time: Option<i64>

The time after which all messages sent to the chat will be automatically deleted; in seconds.

photo: Option<ChatPhoto>

Chat photo.

pinned_message: Option<Message>

The most recent pinned message (by sending date).

profile_accent_color: Option<ProfileAccentColor>

Identifier of the accent color for the chat’s profile background.

profile_background_custom_emoji_id: Option<String>

Custom emoji identifier of the emoji chosen by the chat for its profile background.

active_usernames: Option<Vec<String>>

The list of all active chat usernames; for private chats, supergroups and channels.

username: Option<String>

Username, for private chats, supergroups and channels.

description: Option<String>

Description, for supergroups, channels and group chats.

invite_link: Option<String>

Primary invite link, for supergroups, channels and group chats.

title: Option<String>

Title, for supergroups, channels and group chats.

can_send_paid_media: Option<bool>

Whether paid media messages can be sent or forwarded to the channel chat.

The field is available only for channel chats.

linked_chat_id: Option<i64>

Unique identifier for the linked chat, i.e. the discussion group identifier for a channel and vice versa; for supergroups and channel chats.

bio: Option<String>

Bio of the other party in a private chat.

birthdate: Option<Birthdate>

For private chats, the date of birth of the user.

business_intro: Option<BusinessIntro>

For private chats with business accounts, the intro of the business.

business_location: Option<BusinessLocation>

For private chats with business accounts, the location of the business.

business_opening_hours: Option<BusinessOpeningHours>

For private chats with business accounts, the opening hours of the business.

emoji_status_custom_emoji_id: Option<String>

Custom emoji identifier of the emoji status of the chat or the other party in a private chat.

emoji_status_expiration_date: Option<i64>

Expiration date of the emoji status of the chat or the other party in a private chat, in Unix time, if any.

first_name: Option<String>

First name of the other party in a private chat.

has_private_forwards: Option<bool>

Whether privacy settings of the other party in the private chat allows to use tg://user?id=<user_id> links only in chats with the user.

has_restricted_voice_and_video_messages: Option<bool>

Whether the privacy settings of the other party restrict sending voice and video note messages in the private chat.

last_name: Option<String>

Last name of the other party in a private chat.

personal_chat: Option<Chat>

For private chats, the personal channel of the user.

can_set_sticker_set: Option<bool>

Whether the bot can change the group sticker set.

custom_emoji_sticker_set_name: Option<String>

For supergroups, the name of the group’s custom emoji sticker set. Custom emoji from this set can be used by all users and bots in the group.

has_aggressive_anti_spam_enabled: Option<bool>

Whether aggressive anti-spam checks are enabled in the supergroup. The field is only available to chat administrators.

is_direct_messages: Option<bool>

Whether the chat is the direct messages chat of a channel.

is_forum: Option<bool>

Whether the supergroup chat is a forum (has topics enabled).

join_to_send_messages: Option<bool>

Whether users need to join the supergroup before they can send messages.

join_by_request: Option<bool>

Whether all users directly joining the supergroup need to be approved by supergroup administrators.

location: Option<ChatLocation>

For supergroups, the location to which the supergroup is connected.

parent_chat: Option<Chat>

Information about the corresponding channel chat; for direct messages chats only.

slow_mode_delay: Option<i64>

For supergroups, the minimum allowed delay between consecutive messages sent by each unprivileged user; in seconds.

sticker_set_name: Option<String>

For supergroups, name of the group sticker set.

unrestrict_boost_count: Option<i64>

For supergroups, the minimum number of boosts that a non-administrator user needs to add in order to ignore slow mode and chat permissions.

permissions: Option<ChatPermissions>

Default chat member permissions, for groups and supergroups.

accepted_gift_types: Option<AcceptedGiftTypes>

Information about types of gifts that are accepted by the chat or by the corresponding user for private chats.