Struct carapax::types::InlineKeyboardButton

pub struct InlineKeyboardButton { /* fields omitted */ }

Button of an inline keyboard

You must use exactly one of the optional fields

Methods

impl InlineKeyboardButton

pub fn with_url<S>(text: S, url: S) -> InlineKeyboardButton where
S: Into<String>,

HTTP or tg:// url to be opened when button is pressed

pub fn with_callback_data<S>(text: S, callback_data: S) -> InlineKeyboardButton where
S: Into<String>,

Data to be sent in a callback query to the bot when button is pressed, 1-64 bytes

pub fn with_callback_data_struct<S, D>(
text: S,
callback_data: &D
) -> Result<InlineKeyboardButton, InlineKeyboardError> where
D: Serialize,
S: Into<String>,

Same as with_callback_data, but takes a serializable type

Data will be serialized using serde_json

pub fn with_switch_inline_query<S>(
text: S,
switch_inline_query: S
) -> InlineKeyboardButton where
S: Into<String>,

Pressing the button will prompt the user to select one of their chats, open that chat and insert the bot‘s username and the specified inline query in the input field

Can be empty, in which case just the bot’s username will be inserted

Note: This offers an easy way for users to start using your bot in inline mode when they are currently in a private chat with it

Especially useful when combined with switch_pm… actions – in this case the user will be automatically returned to the chat they switched from, skipping the chat selection screen

pub fn with_switch_inline_query_current_chat<S>(
text: S,
switch_inline_query_current_chat: S
) -> InlineKeyboardButton where
S: Into<String>,

If set, pressing the button will insert the bot‘s username and the specified inline query in the current chat's input field

Can be empty, in which case only the bot’s username will be inserted This offers a quick way for the user to open your bot in inline mode in the same chat – good for selecting something from multiple options

pub fn with_callback_game<S>(text: S) -> InlineKeyboardButton where
S: Into<String>,

Description of the game that will be launched when the user presses the button

NOTE: This type of button must always be the first button in the first row

pub fn with_pay<S>(text: S) -> InlineKeyboardButton where
S: Into<String>,

Send a Pay button

NOTE: This type of button must always be the first button in the first row

pub fn with_login_url<T, U>(text: T, login_url: U) -> InlineKeyboardButton where
T: Into<String>,
U: Into<LoginUrl>,

An HTTP URL used to automatically authorize the user

Can be used as a replacement for the Telegram Login Widget