Struct carapax::types::Location

pub struct Location {
    pub latitude: f32,
    pub longitude: f32,
    pub heading: Option<i64>,
    pub horizontal_accuracy: Option<f32>,
    pub live_period: Option<i64>,
    pub proximity_alert_radius: Option<i64>,
}

Represents a point on a map.

Fields
latitude: f32

Latitude as defined by sender.

longitude: f32

Longitude as defined by sender.

heading: Option<i64>

The direction in which user is moving; in degrees; 1-360.

For active live locations only.

horizontal_accuracy: Option<f32>

The radius of uncertainty for the location, measured in meters; 0-1500.

live_period: Option<i64>

Time relative to the message sending date, during which the location can be updated, in seconds.

For active live locations only.

proximity_alert_radius: Option<i64>

Maximum distance for proximity alerts about approaching another chat member; in meters.

For sent live locations only.

Implementations

impl Location

pub fn new(latitude: f32, longitude: f32) -> Location

Creates a new Location.

Arguments

latitude - Latitude.

longitude - Longitude.

pub fn with_heading(self, value: i64) -> Location

Sets a new heading.

Arguments

value - A direction in which the user is moving; in degrees; 1-360.

pub fn with_horizontal_accuracy(self, value: f32) -> Location

Sets a new horizontal accuracy.

Arguments

value - A radius of uncertainty for the location; in meters; 0-1500.

pub fn with_live_period(self, value: i64) -> Location

Sets a new live period.

Arguments

value - Period in seconds for which the location can be updated; 60-86400.

pub fn with_proximity_alert_radius(self, value: i64) -> Location

Sets a new proximity alert radius.

Arguments

value - A maximum distance for proximity alerts about approaching another chat member; in meters; 1-100000.