Struct carapax::types::PrivateChat

pub struct PrivateChat {

    pub id: ChatPeerId,
    pub first_name: String,
    pub accent_color: Option<AccentColor>,
    pub active_usernames: Option<Vec<String>>,
    pub background_custom_emoji_id: Option<String>,
    pub bio: Option<String>,
    pub emoji_status_custom_emoji_id: Option<String>,
    pub emoji_status_expiration_date: Option<i64>,
    pub has_private_forwards: Option<bool>,
    pub has_restricted_voice_and_video_messages: Option<bool>,
    pub last_name: Option<String>,
    pub message_auto_delete_time: Option<i64>,
    pub photo: Option<ChatPhoto>,
    pub pinned_message: Option<Box<Message>>,
    pub profile_accent_color: Option<ProfileAccentColor>,
    pub profile_background_custom_emoji_id: Option<String>,
    pub username: Option<ChatUsername>,

}

Represents a private chat.

Fields
id: ChatPeerId

Unique identifier of the chat.

first_name: String

First name of the other party.

accent_color: Option<AccentColor>

Identifier of the accent color for the chat name and backgrounds of the chat photo, reply header, and link preview.

Returned only in crate::types::GetChat.

active_usernames: Option<Vec<String>>

List of all active usernames of the other party.

Returned only in crate::types::GetChat.

background_custom_emoji_id: Option<String>

Custom emoji identifier of emoji chosen by the chat for the reply header and link preview background.

Returned only in crate::types::GetChat.

bio: Option<String>

Bio of the other party.

Returned only in crate::types::GetChat.

emoji_status_custom_emoji_id: Option<String>

Custom emoji identifier of emoji status of the other party.

Returned only in crate::types::GetChat.

emoji_status_expiration_date: Option<i64>

Expiration date of the emoji status of the other party in a private chat in Unix time, if any.

Returned only in crate::types::GetChat.

has_private_forwards: Option<bool>

Indicates whether privacy settings of the other party in the private chat allows to use tg://user?id=<user_id> links only in chats with the user.

Returned only in crate::types::GetChat.

has_restricted_voice_and_video_messages: Option<bool>

Indicates whether privacy settings of the other party restrict sending voice and video note messages.

Returned only in crate::types::GetChat.

last_name: Option<String>

Last name of the other party.

message_auto_delete_time: Option<i64>

The time after which all messages sent to the chat will be automatically deleted; in seconds.

Returned only in crate::types::GetChat.

photo: Option<ChatPhoto>

Photo of the other party.

Returned only in crate::types::GetChat.

pinned_message: Option<Box<Message>>

Latest pinned message.

Returned only in crate::types::GetChat.

profile_accent_color: Option<ProfileAccentColor>

Identifier of the accent color for the chat’s profile background.

Returned only in crate::types::GetChat.

profile_background_custom_emoji_id: Option<String>

Custom emoji identifier of the emoji chosen by the chat for its profile background.

Returned only in crate::types::GetChat.

username: Option<ChatUsername>

Username of the target chat.

Implementations

impl PrivateChat

pub fn new<A, B>(id: A, first_name: B) -> PrivateChat
where A: Into<ChatPeerId>, B: Into<String>,

Creates a new PrivateChat.

Arguments

id - Unique identifier of the target chat.

first_name - First name of the other party.

pub fn with_accent_color(self, value: AccentColor) -> PrivateChat

Sets a new accent color.

Arguments

value - Accent color for the chat.

pub fn with_active_usernames<A, B>(self, value: A) -> PrivateChat
where A: IntoIterator<Item = B>, B: Into<String>,

Sets a new list of active usernames.

Arguments

value - Active usernames.

pub fn with_background_custom_emoji_id<T>(self, value: T) -> PrivateChat
where T: Into<String>,

Sets a new custom emoji identifier for the message background.

Arguments

value - Custom emoji identifier of emoji chosen by the chat for the reply header and link preview background.

pub fn with_bio<T>(self, value: T) -> PrivateChat
where T: Into<String>,

Sets a new bio.

Arguments

value - Bio.

pub fn with_emoji_status_custom_emoji_id<T>(self, value: T) -> PrivateChat
where T: Into<String>,

Sets a new custom emoji identifier of emoji status.

Arguments

value - Emoji identifier.

pub fn with_emoji_status_expiration_date(self, value: i64) -> PrivateChat

Sets a new emoji status expiration date.

Arguments

value - Unix timestamp; in seconds.

pub fn with_has_private_forwards(self, value: bool) -> PrivateChat

Sets a new value for a has_private_forwards flag.

Arguments

value - Indicates whether privacy settings of the other party in the private chat allows to use tg://user?id=<user_id> links only in chats with the user.

pub fn with_has_restricted_voice_and_video_messages( self, value: bool ) -> PrivateChat

Sets a new value for a has_restricted_voice_and_video_messages flag.

Arguments

value - Indicates whether privacy settings of the other party restrict sending voice and video note messages.

pub fn with_last_name<T>(self, value: T) -> PrivateChat
where T: Into<String>,

Sets a new last name.

Arguments

value - Last name.

pub fn with_message_auto_delete_time(self, value: i64) -> PrivateChat

Sets a new message auto-delete time.

Arguments

value - Time in seconds.

pub fn with_photo(self, value: ChatPhoto) -> PrivateChat

Sets a new photo.

Arguments

value - Chat photo.

pub fn with_pinned_message(self, value: Message) -> PrivateChat

Sets a new pinned message.

Arguments

value - Pinned message.

pub fn with_profile_accent_color(self, value: ProfileAccentColor) -> PrivateChat

Sets a new profile accent color.

Arguments

value - Accent color for the chat’s profile background.

pub fn with_profile_background_custom_emoji_id<T>(self, value: T) -> PrivateChat
where T: Into<String>,

Sets a new custom emoji identifer for the chat’s profile background.

Arguments

value - Custom emoji identifier of the emoji chosen by the chat for its profile background.

pub fn with_username<T>(self, value: T) -> PrivateChat
where T: Into<ChatUsername>,

Sets a new username.

Arguments

value - Username.