carapax::types

Struct RestrictChatMember

pub struct RestrictChatMember { /* private fields */ }

Restricts a user in a supergroup.

The bot must be an administrator in the supergroup for this to work and must have the appropriate admin rights.

Implementations

impl RestrictChatMember

pub fn new<T>(chat_id: T, user_id: i64) -> RestrictChatMember
where T: Into<ChatId>,

Creates a new RestrictChatMember.

Arguments

chat_id - Unique identifier for the target chat.

user_id - Unique identifier of the target user.

pub fn allow_all(self) -> RestrictChatMember

Allows everything.

pub fn restrict_all(self) -> RestrictChatMember

Restricts everything.

pub fn with_permissions(self, value: ChatPermissions) -> RestrictChatMember

Replaces current permissions with the new one.

Arguments

value - The new permissions.

pub fn with_until_date(self, value: i64) -> RestrictChatMember

Sets a new until date.

Arguments

value - The date when restrictions will be lifted for the user, unix time.

If user is restricted for more than 366 days or less than 30 seconds from the current time, they are considered to be restricted forever.

pub fn with_use_independent_chat_permissions( self, value: bool, ) -> RestrictChatMember

Sets a new value for the use_independent_chat_permissions flag.

Arguments

value - Indicates whether the chat permissions are set independently.

If false, the can_send_other_messages and can_add_web_page_previews permissions will imply the can_send_messages, can_send_audios, can_send_documents, can_send_photos, can_send_videos, can_send_video_notes, and can_send_voice_notes permissions; the can_send_polls permission will imply the can_send_messages permission.