carapax::types

Struct SendAnimation

pub struct SendAnimation { /* private fields */ }

Sends an animation file (GIF or H.264/MPEG-4 AVC video without sound).

Bots can currently send animation files of up to 50 MB in size, this limit may be changed in the future.

Implementations

impl SendAnimation

pub fn new<A, B>(animation: A, chat_id: B) -> SendAnimation
where A: Into<InputFile>, B: Into<ChatId>,

Creates a new SendAnimation.

Arguments

chat_id - Unique identifier of the target chat.

animation - Animation to send.

pub fn with_allow_paid_broadcast(self, value: bool) -> SendAnimation

Sets a new value for the allow_paid_broadcast flag.

Arguments

value - Whether to allow up to 1000 messages per second, ignoring broadcasting limits for a fee of 0.1 Telegram Stars per message. The relevant Stars will be withdrawn from the bot’s balance.

pub fn with_business_connection_id<T>(self, value: T) -> SendAnimation
where T: Into<String>,

Sets a new business connection ID.

Arguments

value - Unique identifier of the business connection.

pub fn with_caption<T>(self, value: T) -> SendAnimation
where T: Into<String>,

Sets a new caption.

Arguments

value - Caption; 0-1024 characters.

May also be used when resending animation by file_id.

pub fn with_caption_entities<T>( self, value: T, ) -> Result<SendAnimation, TextEntityError>
where T: IntoIterator<Item = TextEntity>,

Sets a new list of caption entities.

Arguments

value - The list of special entities that appear in the caption.

Caption parse mode will be set to None when this method is called.

pub fn with_caption_parse_mode(self, value: ParseMode) -> SendAnimation

Sets a new caption parse mode.

Arguments

value - Parse mode.

Caption entities will be set to None when this method is called.

pub fn with_direct_messages_topic_id(self, value: i64) -> SendAnimation

Sets a new direct messages topic ID

value - Identifier of the direct messages topic to which the message will be sent.

Required if the message is sent to a direct messages chat.

pub fn with_disable_notification(self, value: bool) -> SendAnimation

Sets a new value for the disable_notification flag.

Arguments

value - Indicates whether to send the message silently or not; a user will receive a notification without sound.

pub fn with_duration(self, value: i64) -> SendAnimation

Sets a new duration.

Arguments

value - Duration in seconds.

pub fn with_has_spoiler(self, value: bool) -> SendAnimation

Sets a new value for the has_spoiler flag.

Arguments

value - Indicates whether to cover with a spoiler animation.

pub fn with_height(self, value: i64) -> SendAnimation

Sets a new height.

Arguments

value - Height.

pub fn with_message_effect_id<T>(self, value: T) -> SendAnimation
where T: Into<String>,

Sets a new message effect ID.

Arguments

value - Unique identifier of the message effect to be added to the message; for private chats only.

pub fn with_message_thread_id(self, value: i64) -> SendAnimation

Sets a new message thread ID.

Arguments

value - Unique identifier of the target message thread; supergroups only.

pub fn with_protect_content(self, value: bool) -> SendAnimation

Sets a new value for the protect_content flag.

Arguments

value - Indicates whether to protect the contents of the sent message from forwarding and saving.

pub fn with_reply_markup<T>( self, value: T, ) -> Result<SendAnimation, ReplyMarkupError>
where T: Into<ReplyMarkup>,

Sets a new reply markup.

Arguments

value - Reply markup.

pub fn with_reply_parameters( self, value: ReplyParameters, ) -> Result<SendAnimation, ReplyParametersError>

Sets new reply parameters.

Arguments

value - Description of the message to reply to.

pub fn with_show_caption_above_media(self, value: bool) -> SendAnimation

Sets a new value for the show_caption_above_media flag.

Arguments

value - Whether the caption must be shown above the message media.

pub fn with_suggested_post_parameters( self, value: &SuggestedPostParameters, ) -> Result<SendAnimation, SuggestedPostParametersError>

Sets a new suggested post parameters.

Arguments

value - An object containing the parameters of the suggested post to send.

For direct messages chats only.

If the message is sent as a reply to another suggested post, then that suggested post is automatically declined.

pub fn with_thumbnail<T>( self, value: T, ) -> Result<SendAnimation, SendAnimationError>
where T: Into<InputFile>,

Sets a new thumbnail.

Arguments

value - Thumbnail.

The thumbnail should be in JPEG format and less than 200 kB in size. A thumbnail‘s width and height should not exceed 320. Ignored if the file is not uploaded using multipart/form-data. Thumbnails can’t be reused and can be only uploaded as a new file.

pub fn with_width(self, value: i64) -> SendAnimation

Sets a new width.

Arguments

value - Width.