carapax::types

Struct Sticker

pub struct Sticker {

    pub file_id: String,
    pub file_unique_id: String,
    pub height: i64,
    pub is_animated: bool,
    pub is_video: bool,
    pub sticker_type: StickerType,
    pub width: i64,
    pub custom_emoji_id: Option<String>,
    pub emoji: Option<String>,
    pub file_size: Option<i64>,
    pub mask_position: Option<MaskPosition>,
    pub needs_repainting: Option<bool>,
    pub premium_animation: Option<File>,
    pub set_name: Option<String>,
    pub thumbnail: Option<PhotoSize>,

}

Represents a sticker.

Fields
file_id: String

Identifier of the file.

Can be used to download or reuse the file.

file_unique_id: String

Unique identifier of the file.

It is supposed to be the same over time and for different bots. Can’t be used to download or reuse the file.

height: i64

Sticker height.

is_animated: bool

Indicates whether the sticker is animated.

is_video: bool

Indicates whether the sticker is a video sticker.

sticker_type: StickerType

Type of the sticker.

The type of the sticker is independent from its format, which is determined by the fields is_animated and is_video.

width: i64

Sticker width.

custom_emoji_id: Option<String>

For custom emoji stickers, unique identifier of the custom emoji.

emoji: Option<String>

Emoji associated with the sticker.

file_size: Option<i64>

File size in bytes.

mask_position: Option<MaskPosition>

For mask stickers, the position where the mask should be placed.

needs_repainting: Option<bool>

Indicates whether the sticker must be repainted to a text color in messages, the color of the Telegram Premium badge in emoji status, white color on chat photos, or another appropriate color in other places.

premium_animation: Option<File>

For premium regular stickers, premium animation for the sticker.

set_name: Option<String>

Name of the sticker set to which the sticker belongs.

thumbnail: Option<PhotoSize>

Sticker thumbnail in the WEBP or JPEG format.

Implementations

impl Sticker

pub fn new<A, B>( file_id: A, file_unique_id: B, sticker_type: StickerType, height: i64, width: i64, ) -> Sticker
where A: Into<String>, B: Into<String>,

Creates a new Sticker.

Arguments

file_id - Identifier for the file.

file_unique_id - Unique identifier for the file.

sticker_type - Type of the sticker.

height - Sticker height.

width - Sticker width.

pub fn with_is_animated(self, value: bool) -> Sticker

Sets a new value for the is_animated flag.

Arguments

value - Indicates whether the sticker is animated.

pub fn with_is_video(self, value: bool) -> Sticker

Sets a new value for the is_video flag.

Arguments

value - Indicates whether the sticker is a video sticker.

pub fn with_custom_emoji_id<T>(self, value: T) -> Sticker
where T: Into<String>,

Sets a new custom emoji ID.

Arguments

value - Custom emoji ID.

pub fn with_emoji<T>(self, value: T) -> Sticker
where T: Into<String>,

Sets a new emoji.

Arguments

value - Emoji.

pub fn with_file_size(self, value: i64) -> Sticker

Sets a new file size.

Arguments

value - File size in bytes.

pub fn with_mask_position(self, value: MaskPosition) -> Sticker

Sets a new mask position.

Arguments

value - Mask position.

pub fn with_needs_repainting(self, value: bool) -> Sticker

Sets a new value for the needs_repainting flag.

Arguments

value - Value of the flag.

pub fn with_premium_animation(self, value: File) -> Sticker

Sets a new premium animation.

Arguments

value - Premium animation.

pub fn with_set_name<T>(self, value: T) -> Sticker
where T: Into<String>,

Sets a new sticker set name.

Arguments

value - Name of a sticker set.

pub fn with_thumbnail(self, value: PhotoSize) -> Sticker

Sets a new thumbnail.

Arguments

value - Thumbnail.