carapax::access

Trait AccessExt

pub trait AccessExt<P, HI>: Sized {
    // Provided method
    fn with_access_policy(
        self,
        policy: P,
    ) -> Predicate<AccessPredicate<P>, HandlerInput, Self, HI> { ... }
}

Available on crate feature access only.

Provides a shortcut for wrapping a Handler by an AccessPredicate.

Provided Methods

fn with_access_policy( self, policy: P, ) -> Predicate<AccessPredicate<P>, HandlerInput, Self, HI>

Shortcut to wrap a Handler with an access predicate.

Example: let handler = handler.access(policy).

Arguments

policy - A crate::access::AccessPolicy.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.