carapax

Module ratelimit

Available on crate feature ratelimit only.

Ratelimit support

Macros

nonzero

Create non-zero values from constant literals easily.

Structs

DirectRateLimitPredicate

A predicate with a direct rate limiter.

Jitter

An interval specification for deviating from the nominal wait time.

KeyChat

Represents a rate limit key for a chat.

KeyChatUser

Represents a rate limit key for a user in a chat.

KeyUser

Represents a rate limit key for a user.

KeyedRateLimitPredicate

A predicate with keyed rate limiter.

MethodDiscard

Discards updates when the rate limit is reached.

MethodWait

Allows update to pass as soon as the rate limiter allows it.

NoJitter

Wait without jitter.

Quota

A rate-limiting quota.

Traits

Key

Represents a key for a keyed rate limiter.